+++ testing/reports/logs/clover.xml
@@ -1,61 +1,61 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<coverage generated="1554508211">
-  <project timestamp="1554508211">
+<coverage generated="1554512494">
+  <project timestamp="1554512494">
     <file name="D:\Documents\Discuss\Vani Ahmad Ramadhan\tesis\apriori.php">
-      <line num="4" type="stmt" count="6"/>
-      <line num="6" type="stmt" count="6"/>
-      <line num="7" type="stmt" count="6"/>
-      <line num="8" type="stmt" count="6"/>
-      <line num="10" type="stmt" count="6"/>
-      <line num="11" type="stmt" count="6"/>
-      <line num="14" type="stmt" count="6"/>
-      <line num="15" type="stmt" count="6"/>
-      <line num="19" type="stmt" count="6"/>
-      <line num="24" type="stmt" count="8"/>
-      <line num="26" type="stmt" count="8"/>
-      <line num="27" type="stmt" count="8"/>
-      <line num="29" type="stmt" count="8"/>
-      <line num="30" type="stmt" count="8"/>
-      <line num="31" type="stmt" count="8"/>
-      <line num="34" type="stmt" count="8"/>
-      <line num="38" type="stmt" count="8"/>
-      <line num="40" type="stmt" count="8"/>
-      <line num="45" type="stmt" count="6"/>
-      <line num="46" type="stmt" count="6"/>
-      <line num="49" type="stmt" count="6"/>
-      <line num="54" type="stmt" count="6"/>
-      <line num="55" type="stmt" count="6"/>
-      <line num="57" type="stmt" count="6"/>
-      <line num="58" type="stmt" count="6"/>
-      <line num="59" type="stmt" count="6"/>
-      <line num="61" type="stmt" count="6"/>
-      <line num="62" type="stmt" count="6"/>
-      <line num="63" type="stmt" count="6"/>
-      <line num="64" type="stmt" count="6"/>
+      <line num="4" type="stmt" count="9"/>
+      <line num="6" type="stmt" count="9"/>
+      <line num="7" type="stmt" count="8"/>
+      <line num="8" type="stmt" count="8"/>
+      <line num="10" type="stmt" count="8"/>
+      <line num="11" type="stmt" count="8"/>
+      <line num="14" type="stmt" count="8"/>
+      <line num="15" type="stmt" count="8"/>
+      <line num="19" type="stmt" count="9"/>
+      <line num="24" type="stmt" count="12"/>
+      <line num="26" type="stmt" count="12"/>
+      <line num="27" type="stmt" count="11"/>
+      <line num="29" type="stmt" count="11"/>
+      <line num="30" type="stmt" count="11"/>
+      <line num="31" type="stmt" count="11"/>
+      <line num="34" type="stmt" count="11"/>
+      <line num="38" type="stmt" count="12"/>
+      <line num="40" type="stmt" count="12"/>
+      <line num="45" type="stmt" count="9"/>
+      <line num="46" type="stmt" count="9"/>
+      <line num="49" type="stmt" count="9"/>
+      <line num="54" type="stmt" count="10"/>
+      <line num="55" type="stmt" count="10"/>
+      <line num="57" type="stmt" count="10"/>
+      <line num="58" type="stmt" count="9"/>
+      <line num="59" type="stmt" count="9"/>
+      <line num="61" type="stmt" count="9"/>
+      <line num="62" type="stmt" count="7"/>
+      <line num="63" type="stmt" count="7"/>
+      <line num="64" type="stmt" count="7"/>
       <line num="65" type="stmt" count="2"/>
-      <line num="68" type="stmt" count="6"/>
-      <line num="72" type="stmt" count="6"/>
-      <line num="74" type="stmt" count="6"/>
-      <line num="75" type="stmt" count="6"/>
-      <line num="77" type="stmt" count="6"/>
-      <line num="78" type="stmt" count="4"/>
-      <line num="79" type="stmt" count="4"/>
-      <line num="82" type="stmt" count="3"/>
-      <line num="83" type="stmt" count="3"/>
-      <line num="89" type="stmt" count="6"/>
-      <line num="95" type="stmt" count="8"/>
-      <line num="96" type="stmt" count="8"/>
-      <line num="102" type="stmt" count="6"/>
-      <line num="103" type="stmt" count="6"/>
-      <line num="106" type="stmt" count="6"/>
-      <line num="107" type="stmt" count="6"/>
-      <line num="108" type="stmt" count="6"/>
-      <line num="110" type="stmt" count="6"/>
-      <line num="111" type="stmt" count="6"/>
-      <line num="114" type="stmt" count="6"/>
-      <line num="115" type="stmt" count="6"/>
-      <line num="119" type="stmt" count="6"/>
-      <line num="120" type="stmt" count="6"/>
+      <line num="68" type="stmt" count="7"/>
+      <line num="72" type="stmt" count="8"/>
+      <line num="74" type="stmt" count="8"/>
+      <line num="75" type="stmt" count="8"/>
+      <line num="77" type="stmt" count="8"/>
+      <line num="78" type="stmt" count="6"/>
+      <line num="79" type="stmt" count="6"/>
+      <line num="82" type="stmt" count="5"/>
+      <line num="83" type="stmt" count="5"/>
+      <line num="89" type="stmt" count="10"/>
+      <line num="95" type="stmt" count="9"/>
+      <line num="96" type="stmt" count="10"/>
+      <line num="102" type="stmt" count="9"/>
+      <line num="103" type="stmt" count="9"/>
+      <line num="106" type="stmt" count="9"/>
+      <line num="107" type="stmt" count="7"/>
+      <line num="108" type="stmt" count="7"/>
+      <line num="110" type="stmt" count="7"/>
+      <line num="111" type="stmt" count="7"/>
+      <line num="114" type="stmt" count="7"/>
+      <line num="115" type="stmt" count="7"/>
+      <line num="119" type="stmt" count="9"/>
+      <line num="120" type="stmt" count="9"/>
       <line num="125" type="stmt" count="6"/>
       <line num="127" type="stmt" count="6"/>
       <line num="128" type="stmt" count="6"/>
